+++ cmd/server.go
@@ -132,7 +132,7 @@ var serverCmd = &cobra.Command{
 		logger.Info("initialized service")
 
 		logger.Info("initializing http handlers")
-		var handler = tshttp.Handler(svc, jwtSecret, logger.With("component", "HTTP"))
+		var handler = tshttp.Handler(svc, jwtSecret, logger.With("component", "HTTP"), db, cfg.DBSystem)
 		logger.Info("initialized http handlers")
 
 		reg := prometheus.NewRegistry()